mejorada vista perfil e implementado sistema de apertura de incidencias

+++ routes/web.php
@@ -9,6 +9,8 @@ use App\Http\Controllers\userController;
 use App\Http\Controllers\ValoracionController;
 use App\Http\Controllers\ContactController;
 use App\Http\Controllers\ItemController;
+use App\Http\Controllers\PagoController;
+use App\Http\Controllers\IncidenciaController;
 
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
@@ -27,24 +29,24 @@ Route::controller(IndexController::class)->group(function(){
 
 
 
-Route::get('/alquileres/arrendatario/{id_usuario}', [AlquilerController::class, 'alquileresPorArrendatario']);
-Route::get('/alquileres/arrendador/{id_usuario}', [AlquilerController::class, 'alquileresPorArrendador']);
+Route::get('/alquileres/arrendatario/{id_usuario}', [AlquilerController::class, 'alquileresPorArrendatario'])->middleware('auth');
+Route::get('/alquileres/arrendador/{id_usuario}', [AlquilerController::class, 'alquileresPorArrendador'])->middleware('auth');
 Auth::routes();
 
 Route::get('/categorias', [CategoriaController::class,'index'])->name('categorias.index');
-Route::get('/perfil', [UserController::class, 'show'])->name('perfil');
+Route::get('/perfil', [UserController::class, 'show'])->name('perfil')->middleware('auth');
 
 
 Route::get('/subcategorias/{categoria_id}', [CategoriaController::class, 'getSubcategorias']);
 
-Route::get('/productos.create', [ProductoController::class, 'create'])->name('productos.create');
+Route::get('/productos.create', [ProductoController::class, 'create'])->name('productos.create')->middleware('auth');
 Route::get('/productos/buscar', [ProductoController::class, 'buscar'])->name('productos.buscar');
-Route::post('/productos/store', [ProductoController::class, 'store'])->name('productos.store'); 
+Route::post('/productos/store', [ProductoController::class, 'store'])->name('productos.store')->middleware('auth');
 Route::get('/productos/index', [ProductoController::class, 'index'])->name('productos.index'); 
-Route::post('/productos/actualizar/{producto}', [ProductoController::class, 'actualizar'])->name('productos.actualizar'); 
-Route::post('/productos/edit', [ProductoController::class, 'edit'])->name('productos.edit'); 
-Route::get('/productos/misproductos', [ProductoController::class, 'misproductos'])->name('productos.misproductos'); 
-Route::delete('/productos/{producto}', [ProductoController::class, 'destroy'])->name('productos.eliminar');
+Route::post('/productos/actualizar/{producto}', [ProductoController::class, 'actualizar'])->name('productos.actualizar')->middleware('auth');
+Route::post('/productos/edit', [ProductoController::class, 'edit'])->name('productos.edit')->middleware('auth'); 
+Route::get('/productos/misproductos', [ProductoController::class, 'misproductos'])->name('productos.misproductos')->middleware('auth'); 
+Route::delete('/productos/{producto}', [ProductoController::class, 'destroy'])->name('productos.eliminar')->middleware('auth');
 Route::get('/productos/verproducto/{producto}', [ProductoController::class, 'verproducto'])->name('productos.verproducto');
 Route::get('/productos/novedades', [ProductoController::class, 'novedades'])->name('productos.novedades');
 Route::get('/productos/ofertas', [ProductoController::class, 'ofertas'])->name('productos.ofertas');
@@ -65,20 +67,24 @@ Route::view('/general/politica', 'general.politica');
 
 Route::get('/subcategoria/{id}', [SubcategoriaController::class, 'index'])->name('subcategoria.index');
 
-Route::post('/productos/{producto}/reservar', [ProductoController::class, 'actualizarReserva'])->name('productos.actualizarReserva');
-Route::get('/productos/{producto}/alquileres', [ProductoController::class, 'obtenerAlquileres'])->name('productos.obtenerAlquileres');
+Route::post('/productos/{producto}/reservar', [ProductoController::class, 'actualizarReserva'])->name('productos.actualizarReserva')->middleware('auth');
+Route::get('/productos/{producto}/alquileres', [ProductoController::class, 'obtenerAlquileres'])->name('productos.obtenerAlquileres')->middleware('auth');
 
 
-Route::get('/alquileres/{id}/edit/{id_producto}', [AlquilerController::class, 'edit'])->name('alquileres.edit');
-Route::delete('/alquileres/{id}/cancel', [AlquilerController::class, 'cancelar'])->name('alquileres.cancel');
-Route::put('/alquileres/{id}', [AlquilerController::class, 'update'])->name('alquileres.update');
+Route::get('/alquileres/{id}/edit/{id_producto}', [AlquilerController::class, 'edit'])->name('alquileres.edit')->middleware('auth');
+Route::delete('/alquileres/{id}/cancel', [AlquilerController::class, 'cancelar'])->name('alquileres.cancel')->middleware('auth');
+Route::put('/alquileres/{id}', [AlquilerController::class, 'update'])->name('alquileres.update')->middleware('auth');
 
-Route::post('/valoraciones/guardar', [ValoracionController::class, 'guardar'])->name('valoraciones.guardar');
+Route::post('/valoraciones/guardar', [ValoracionController::class, 'guardar'])->name('valoraciones.guardar')->middleware('auth');
 
-Route::get('/publicar-oferta', [ItemController::class, 'index'])->name('ofertas.index'); // Ver listado
-Route::post('/publicar-oferta', [ItemController::class, 'store'])->name('ofertas.store'); // Crear nuevo item
-Route::delete('/items/{id}', [ItemController::class, 'destroy'])->name('items.destroy');
+Route::get('/publicar-oferta', [ItemController::class, 'index'])->name('ofertas.index')->middleware('auth'); // Ver listado
+Route::post('/publicar-oferta', [ItemController::class, 'store'])->name('ofertas.store')->middleware('auth'); // Crear nuevo item
+Route::delete('/items/{id}', [ItemController::class, 'destroy'])->name('items.destroy')->middleware('auth');
 
 Route::get('/presentacion', function () {
     return view('general.presentacion');
 });
+
+
+Route::get('/pagos/detalles/{id}', [PagoController::class, 'detalles'])->name('pagos.detalles');
+Route::post('/incidencias', [IncidenciaController::class, 'store'])->name('incidencias.store');
